Refactors buildsystem into playerUI and gives it to the player
