+++ Blocktest/Assets/Scripts/PlayerUI.cs
@@ -2,14 +2,19 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Tilemaps;
-/* TODO: Finish seperating the UI elements of BuildSystem.cs to here
+using UnityEngine.UI;
 public class PlayerUI : MonoBehaviour
 {
-    /// The linked build system component.
+
+    /// Reference to main BlockManager script.
+    private BlockManager blockManager;
+    /// Reference to the main BuildSystem script.
     private BuildSystem buildSystem;
 
-    /// Maximum distance at which the player can place blocks
-    [SerializeField] float maxBuildDistance = 3f;
+    /// The ID of the currently selected block.
+    private int currentBlockID = 0;
+    /// The currently selected block.
+    private Block currentBlock;
 
     ///The block placement template object.
     private GameObject blockCursor;
@@ -17,31 +22,54 @@ public class PlayerUI : MonoBehaviour
     private SpriteRenderer currentRenderer;
     /// The block placement template's audio source
     private AudioSource audioSource;
+    /// The sprite to show around the cursor when in destroy mode.
+    [SerializeField] Sprite destroySprite;
+
+    /// Whether build mode is on or not.
+    public bool buildMode = false;
+
+    /// Maximum distance at which the player can place blocks
+    [SerializeField] float maxBuildDistance = 5f;
+    /// Dropdown used for item selection
+    [SerializeField] Dropdown selectionDropdown;
 
     /// Tilemap for foreground objects
     [SerializeField] Tilemap foregroundTilemap;
     ///Tilemap for background (non-dense) objects
     [SerializeField] Tilemap backgroundTilemap;
 
-    // Start is called before the first frame update
+
     void Start()
     {
-        buildSystem = GetComponent<BuildSystem>();
+        GameObject BlockSystem = GameObject.Find("BlockSystem");
+        buildSystem = BlockSystem.GetComponent<BuildSystem>();
+        blockManager = BlockSystem.GetComponent<BlockManager>();
+        InitializeCursor();
     }
 
-    // Update is called once per frame
     void Update()
     {
+        if(Input.GetKeyDown("e"))
+        {
+            ToggleBuild();
+        }
+
         Vector3Int tilePosition = foregroundTilemap.WorldToCell(Camera.main.ScreenToWorldPoint(Input.mousePosition));
         Vector3 worldTilePosition = foregroundTilemap.CellToWorld(tilePosition) + foregroundTilemap.cellSize / 2;
         blockCursor.transform.position = foregroundTilemap.CellToWorld(tilePosition) + foregroundTilemap.cellSize / 2;
 
-        if(buildSystem.buildMode) { //Things to happen in "build mode"
+        if(Vector2.Distance(worldTilePosition, gameObject.transform.position) > maxBuildDistance) {
+            currentRenderer.color = new Color(1f, 0f, 0f, 0.7f);
+            return;
+        }
+
+        if(buildMode) 
+        {
             float scrollInput = Input.GetAxis("Mouse ScrollWheel");
             if(scrollInput != 0)
             {
                 // Change ID by -1 if scroll input is greater than zero, otherwise change ID by +1.
-                buildSystem.CycleBlockSelection(scrollInput > 0 ? -1 : 1);
+                CycleBlockSelection(scrollInput > 0 ? -1 : 1);
             }
 
             currentRenderer.sprite = currentBlock.blockSprite;
@@ -55,11 +83,6 @@ public class PlayerUI : MonoBehaviour
                 }
             }
 
-            if(Vector2.Distance(worldTilePosition, playerObject.transform.position) > maxBuildDistance) {
-                canBuildBackground = false;
-                canBuildForeground = false;
-            }
-
             if(!canBuildForeground) {
                 currentRenderer.color = new Color(1f, 0f, 0f, 0.7f); // Red if you can't build on the foreground
             } else if(!canBuildBackground) {
@@ -68,21 +91,29 @@ public class PlayerUI : MonoBehaviour
                 currentRenderer.color = new Color(0.5f, 1f, 0.5f, 0.7f); // Otherwise, normal coloring
             }
 
-            if (canBuildForeground && Input.GetMouseButton(0))
-            {
-                PlaceBlock(currentBlock, true, Camera.main.ScreenToWorldPoint(Input.mousePosition));
+            if (canBuildForeground && Input.GetMouseButton(0)) {
+                PlayerPlaceBlock(currentBlock, true, Camera.main.ScreenToWorldPoint(Input.mousePosition));
+            } 
+            else if (canBuildBackground && Input.GetMouseButton(1)) {
+                PlayerPlaceBlock(currentBlock, false, Camera.main.ScreenToWorldPoint(Input.mousePosition));
             }
-            if (canBuildBackground && Input.GetMouseButton(1))
-            {
-                PlaceBlock(currentBlock, false, Camera.main.ScreenToWorldPoint(Input.mousePosition));
-            }
-        } else { // Things to happen in "destroy mode"
-            if(Input.GetMouseButton(0)) {
-                buildSystem.BreakBlock(Camera.main.ScreenToWorldPoint(Input.mousePosition));
-            }
+
         }
+        else
+        {
+            if(Input.GetMouseButton(0)) {
+                PlayerBreakBlock(true, Camera.main.ScreenToWorldPoint(Input.mousePosition));
+            } else if(Input.GetMouseButton(1)) {
+                PlayerBreakBlock(false, Camera.main.ScreenToWorldPoint(Input.mousePosition));
+            }
+            
+        }
+
     }
 
+    //
+    // Summary:
+    //      Initialize the cursor block template.
     private void InitializeCursor()
     {
         if(blockCursor) {
@@ -91,10 +122,108 @@ public class PlayerUI : MonoBehaviour
         blockCursor = new GameObject("BlockCursor");
         currentRenderer = blockCursor.AddComponent<SpriteRenderer>();
         audioSource = blockCursor.AddComponent<AudioSource>();
-        if(!buildSystem.buildMode) {
+        if(!buildMode) {
+            currentRenderer.sprite = destroySprite;
+        }
+        if (currentBlock == null) {
+            //Ensure the block ID is valid.
+            if (blockManager.allBlocks[currentBlockID] != null)
+            {
+                currentBlock = blockManager.allBlocks[currentBlockID];
+            }
+        }
+    }
+
+    //
+    // Summary:
+    //      Toggle build mode on and off.
+    public void ToggleBuild()
+    {
+        buildMode = !buildMode;
+        if (blockCursor == null)
+        {
+            InitializeCursor();
+        }
+
+        //Set the current block.
+        if (currentBlock == null)
+        {
+            //Ensure the block ID is valid.
+            if (blockManager.allBlocks[currentBlockID] != null)
+            {
+                currentBlock = blockManager.allBlocks[currentBlockID];
+            }
+        }
+        currentRenderer.color = new Color(1f, 1f, 1f, 1f);
+        if(buildMode) {
+            currentRenderer.sprite = currentBlock.blockSprite;
+        } else {
             currentRenderer.sprite = destroySprite;
         }
     }
 
+    //
+    // Summary:
+    //      Change the player's currently selected block by "cycling" a certain amount of slots.
+    // Parameters:
+    //      slotDelta:
+    //          The amount of slots to "cycle."
+    public void CycleBlockSelection(int slotDelta)
+    {
+        int totalBlocks = blockManager.allBlocks.Length - 1;
+        int newBlockID = currentBlockID + slotDelta;
+        if(newBlockID > totalBlocks) {
+            newBlockID = 0;
+        } else if(newBlockID < 0) {
+            newBlockID = totalBlocks;
+        }
+        ChangeBlockSelection(newBlockID);
+    }
+
+    //
+    // Summary:
+    //      Change the player's currently selected block.
+    // Parameters:
+    //      slot:
+    //          The slot to change to.
+    public void ChangeBlockSelection(int slot)
+    {
+        slot = Mathf.Clamp(slot, 0, blockManager.allBlocks.Length - 1);
+        currentBlockID = slot;
+        currentBlock = blockManager.allBlocks[currentBlockID];
+        selectionDropdown.captionText.text = currentBlock.blockName;
+        if(buildMode) {
+            currentRenderer.sprite = currentBlock.blockSprite;
+        }
+    }
+
+    //
+    // Summary:
+    //      The method called whenever a PLAYER places an object.
+    // Parameters:
+    //      toPlace:
+    //          The block type to place.
+    //      foreground:
+    //          Whether or not the block should be placed in the foreground.
+    //      position:
+    //          The position of the placed block (world coords)
+    private void PlayerPlaceBlock(Block toPlace, bool foreground, Vector2 position)
+    {
+        audioSource.PlayOneShot(toPlace.placeSound);
+        buildSystem.PlaceBlockWorld(toPlace, foreground, position);
+    }
+
+    //
+    // Summary:
+    //      The method called whenever a PLAYER tries to break an object.
+    // Parameters:
+    //      foreground:
+    //          Whether or not the block to be destroyed is in the foreground.
+    //      position:
+    //          The position of the block to destroy (world coords)
+    private void PlayerBreakBlock(bool foreground, Vector2 position)
+    {
+        buildSystem.BreakBlockWorld(foreground, position);
+    }
+
 }
-*/
